pyproject.toml
[tool.uv.sources]
cake = { path = ".", editable = true }

[project]
name = "cake"
version = "0.1.0"
description = "CAKE - Corporate Aggregation & Knowledge Extraction. Extract content from Jira, Confluence, and Google Drive with permissions for RAG ingestion."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.3",
    "google-api-python-client",
    "google-auth",
    "python-dotenv"
]

[project.scripts]
cake = "cake:main"
